refactor to call generic impl in laml_api_impl.js
